requirements.txt
python-dotenv
streamlit
unstructured
tiktoken
langchain-chroma
langchain.groq
langchain.community
protobuf
langchain-huggingface